extract line logic
